Update shell-command.coffee
The original script could return nulls, which the Slack adapter didn't seem to like and it failed to print the msg.send stdout. (The example of host lookup would work flawlessly, but even pwd or ls would return nothing.) This meant that commands that returned no null would print to Slack and commands that returned nulls would not, hence inconsistencies in return from stdout. This version is a little more robust than the original in that it will ignore nulls and not pass them to the adapter.